Update deploy-staging.yaml: set replicaCount based on ENABLE environment variable
This commit updates the deploy-staging.yaml file to set the replicaCount parameter in the helm upgrade command based on the value of the ENABLE environment variable. If ENABLE is 'true', replicaCount will be set to '1'; otherwise, it will be set to '0'. This change ensures that the number of replicas for the deployment is determined dynamically based on the environment configuration.